service done

+++ model/user_quiz_result.go
@@ -6,9 +6,10 @@ type UserQuizResult struct {
 	gorm.Model
 	UserId int
 	User User `gorm:"foreignKey:UserId;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
+	
 	QuizId int
 	Quiz Quiz `gorm:"foreignKey:QuizId;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
-	Score int
+	Score float64
 	TotalQuestions int
 	CorrectAnswers int
 }
